requirements.txt
streamlit
langchain
PyPDF2
sentence-transformers
faiss-cpu
openai
tiktoken
python-dotenv
